ovn: Build the image only once instead
The build for every node is unreliable and the chance for it to
fail increases with every added node. Build the image only once
at the first master and then distribute it to other nodes.

The build process uses commandline SCP to transfer the images
to all nodes, the reason for that is the speed difference.
SFTP was able to transfer with speed of 20Mbps, however SCP is
able to transfer with >800Mbps. This reduces the time to deploy
all images significantly on scale.

Signed-off-by: Ales Musil <amusil@redhat.com>