clustersConfig: Do not recreate the NodeConfig
The NodeConfig is created once for each worker, that results in MAC
being generated when it's not present. The list of workers that
are specified by the batch (--workers/--skip--workers) is list of
NodeConfigs too, however the second initialization would lead to
MAC collisions as the second list would always collide between runs
if the batch shifted.

Make sure the second NodeConfig is actually the same as original one
so the MAC can stay consistent between batches.

Signed-off-by: Ales Musil <amusil@redhat.com>